requirements.txt
python-telegram-bot
openai
requests
streamlit
pandas
plotly
matplotlib
python-dotenv
requests
transformers
torch
safetensors
